🔒 Fix html.security.plaintext-http-link.plaintext-http-link vulnerability in app/views/tutorial/a2.html
The original code contains a link using plaintext HTTP protocol (http://), which is vulnerable to man-in-the-middle attacks where attackers can potentially intercept and modify the traffic between the user and the website. Using HTTPS ensures that the connection is encrypted and the data transfer is secure. Wikipedia supports HTTPS connections, so updating the URL to use HTTPS instead of HTTP provides better security for users clicking the link. This change ensures that users are directed to the secure version of Wikipedia, protecting against potential traffic interception or manipulation.